+++ src/entities/course/api/course-api.ts
@@ -1,14 +1,6 @@
-import {
-  CoursePageResponse,
-  CoursesResponse,
-  CoursesScheduleResponse,
-} from '@/entities/course/types';
-import {
-  API_CONTENT_TYPE_DICTIONARY,
-  API_MAX_INCLUDE_DEPTH,
-  API_OMIT_LINKED_ITEMS_INCLUDE_DEPTH,
-} from '@/shared/constants';
-import { ApiResourceLocale, ApiServices } from '@/shared/types';
+import { CoursesResponse, CoursesScheduleResponse } from '@/entities/course/types';
+import { API_CONTENT_TYPE_DICTIONARY, API_MAX_INCLUDE_DEPTH } from '@/shared/constants';
+import { ApiServices } from '@/shared/types';
 
 export class CourseApi {
   constructor(private readonly services: ApiServices) {}
@@ -36,25 +28,5 @@ export class CourseApi {
 
   public queryCoursesSchedule() {
     return this.services.rest.get<CoursesScheduleResponse>('/app/courses.json');
-  }
-
-  public queryCoursePage(slug: string, locale: ApiResourceLocale = 'en-US') {
-    return this.services.rest.get<CoursePageResponse>('/entries', {
-      params: {
-        'content_type': API_CONTENT_TYPE_DICTIONARY.COURSE_PAGE,
-        'include': API_MAX_INCLUDE_DEPTH,
-        'fields.slug': slug,
-        locale,
-      },
-    });
-  }
-
-  public queryCoursePages() {
-    return this.services.rest.get<CoursePageResponse>('/entries', {
-      params: {
-        content_type: API_CONTENT_TYPE_DICTIONARY.COURSE_PAGE,
-        include: API_OMIT_LINKED_ITEMS_INCLUDE_DEPTH,
-      },
-    });
   }
 }
