+++ src/entities/course/index.ts
@@ -1,5 +1,4 @@
 export type { ApiCoursesIds, Course, CourseItemData, CourseStatus, Section } from './types';
 export { CourseCard } from './ui/course-card/course-card';
 export { CourseItem } from './ui/course-item/course-item';
-export { SectionResolver } from './ui/section-resolver';
 export { courseStore } from './model/store';
