+++ src/views/course/helpers/transform-course-pages.ts
