+++ src/views/course/model/store.ts
@@ -1,0 +1,44 @@
+import { CoursePageResponse } from '@/entities/course/types';
+import { api } from '@/shared/api/api';
+import { prepareContentfulResponse } from '@/shared/helpers/prepare-contentful-response';
+import { ApiResourceLocale } from '@/shared/types';
+import { transformCoursePages } from '@/views/course/helpers/transform-course-pages';
+import { transformCourseSections } from '@/views/course/helpers/transform-course-sections';
+
+class CoursePageStore {
+  public loadCoursePage = async (slug: string, locale: ApiResourceLocale = 'en-US') => {
+    const res = await api.coursePage.queryCoursePage(slug, locale);
+
+    if (res.isSuccess) {
+      const preparedData = prepareContentfulResponse<CoursePageResponse['items']>(res.result);
+
+      const { title = '', sections: coursePageSections, course } = preparedData.at(0)?.fields ?? {};
+      const courseId = course?.sys?.id;
+      const sections = transformCourseSections(coursePageSections);
+
+      if (!courseId) {
+        throw new Error('Course id is not defined.');
+      }
+
+      return {
+        courseId,
+        courseName: title,
+        sections,
+      };
+    }
+
+    throw new Error('Something went wrong fetching course page!');
+  };
+
+  public loadCoursePages = async () => {
+    const res = await api.coursePage.queryCoursePages();
+
+    if (res.isSuccess) {
+      return transformCoursePages(res.result);
+    }
+
+    throw new Error('Something went wrong fetching all course pages!');
+  };
+}
+
+export const coursePageStore = new CoursePageStore();
