+++ src/views/course/section-resolver.tsx
