+++ src/widgets/about-course/helpers/is-about-course-section.ts
