+++ src/widgets/about-course/helpers/transform-grid-items.ts
