+++ src/widgets/learning-path-stages/helpers/is-learning-path-stages-section.ts
