+++ src/widgets/learning-path-stages/helpers/transform-learning-path-stage-item.ts
