+++ src/widgets/learning-path-stages/index.ts
@@ -1,2 +1,5 @@
 export { LearningPathStageItem } from './ui/learning-path-stage-item/learning-path-stage-item';
 export { LearningPathStages } from './ui/learning-path-stages/learning-path-stages';
+export { isLearningPathStagesSection } from './helpers/is-learning-path-stages-section';
+export { transformLearningPathStageItem } from './helpers/transform-learning-path-stage-item';
+export { transformLearningPathStages } from './helpers/transform-learning-path-stages';
