+++ src/widgets/media-text-block/helpers/is-media-text-block-section.ts
