+++ src/widgets/media-text-block/helpers/transform-media-text-block-section.ts
