+++ src/widgets/video-block/helpers/is-video-block-section.ts
