+++ src/widgets/video-block/helpers/transform-video-block-section.ts
