+++ src/widgets/video-block/index.ts
@@ -1,1 +1,3 @@
 export { VideoBlock } from './ui/video-block';
+export { isVideoBlockSection } from './helpers/is-video-block-section';
+export { transformVideoBlockSection } from './helpers/transform-video-block-section';
